multi_agent_interview/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
aiohttp==3.12.15
PyYAML==6.0.3
